+++ peer_web_frontend/src/app/page.tsx
@@ -1,3 +1,3 @@
 export default function Home() {
-  return <>peer front</>
+  return <></>
 }
