+++ peer_web_frontend/styles/global.css
@@ -5,7 +5,7 @@
   }
 }
 
-@media (min-width: 361px) {
+@media (min-width: 481px) {
   .layout {
     display: flex;
     flex-direction: row;
